Update release.yml
Build before testing: Although you’re using the --no-build flag with the dotnet test command (which should be fine if the build step is working correctly), you might want to try removing this flag. This will force the test command to build the test project again before running the tests.